requirements.txt
annotated-types==0.6.0
anyio==3.7.1
Cerberus==1.3.5
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
setuptools==68.2.2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
